revert changes to model

+++ vllm/model_executor/models/llama.py
@@ -153,7 +153,7 @@ class LlamaAttention(nn.Module):
         qkv, _ = self.qkv_proj(hidden_states)
         q, k, v = qkv.split([self.q_size, self.kv_size, self.kv_size], dim=-1)
         q, k = self.rotary_emb(positions, q, k)
-        attn_output = self.attn(q.to(torch.float32), k.to(torch.float32), v.to(torch.float32), kv_cache, attn_metadata).to(torch.bfloat16)
+        attn_output = self.attn(q, k, v, kv_cache, attn_metadata)
         output, _ = self.o_proj(attn_output)
         return output
